refactored schema validation to Schema.groovy
